:recycle: refactoring Dialog to use flexible
